renaming and refactoring - small changes
