+++ App.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import Navigator from './src/navigator/AppNavigation';
+import Navigator from './src/navigation/AppNavigation';
 
 export default class App extends React.Component {
     render() {
