+++ src/navigation/ReduxNavigation.js
